Changed "current" to "current_weather", added a utility.py file for getting coordinates, and removed import tabulate from pip as it didn't serve a function.
